+++ src/main/java/net/helpscout/api/model/customfield/CustomFieldType.java
@@ -1,0 +1,9 @@
+package net.helpscout.api.model.customfield;
+
+public enum CustomFieldType {
+    DROPDOWN,
+    SINGLE_LINE,
+    MULTI_LINE,
+    DATE,
+    NUMBER
+}