+++ src/main/java/net/helpscout/api/model/ref/TeamRef.java
@@ -1,0 +1,5 @@
+package net.helpscout.api.model.ref;
+
+public class TeamRef extends MailboxUserRef {
+    
+}